+++ packages/excalidraw/clients.ts
@@ -2,7 +2,7 @@ import {
   COLOR_CHARCOAL_BLACK,
   COLOR_VOICE_CALL,
   COLOR_WHITE,
-  THEME,
+  isDarkTheme,
   UserIdleState,
 } from "@excalidraw/common";
 
@@ -124,7 +124,7 @@ export const renderRemoteCursors = ({
 
     // TODO remove the dark theme color after we stop inverting canvas colors
     const IS_SPEAKING_COLOR =
-      appState.theme === THEME.DARK ? "#2f6330" : COLOR_VOICE_CALL;
+      isDarkTheme(appState.theme) ? "#2f6330" : COLOR_VOICE_CALL;
 
     const isSpeaking = collaborator?.isSpeaking;